+++ lib/src/util/extensions/ext.dart
@@ -1,3 +1,4 @@
+export 'package:cubes/src/util/extensions/bool_extensions.dart';
 export 'package:cubes/src/util/extensions/build_context_extensions.dart';
 export 'package:cubes/src/util/extensions/observable_value_extensions.dart';
 export 'package:cubes/src/util/extensions/string_extensions.dart';