+++ lib/src/widgets/cube_consumer.dart
@@ -45,7 +45,9 @@ class _CubeConsumerState<C extends Cube> extends State<CubeConsumer>
 
   @override
   void dispose() {
-    if (cubeWidget.dispose?.call(cube) ?? true) cube.dispose();
+    if (cubeWidget.dispose?.call(cube) ?? true) {
+      cube.dispose();
+    }
     cube.removeOnActionListener(_onAction);
     super.dispose();
   }
@@ -59,14 +61,70 @@ class _CubeConsumerState<C extends Cube> extends State<CubeConsumer>
   }
 
   void _onAction(C cube, CubeAction data) {
+    if (data is NavigationAction) {
+      _resolveNavigation(context, data);
+    }
     postFrame(() => cubeWidget.onAction?.call(cube, data));
   }
 
   void _ready(_) {
-    var arguments =
-        widget.arguments ?? ModalRoute.of(context)?.settings.arguments;
-    cube.onReady(arguments);
+    cube.onReady(widget.arguments ?? context.arguments);
   }
 
   CubeConsumer<C> get cubeWidget => (widget as CubeConsumer<C>);
+
+  void _resolveNavigation(BuildContext context, NavigationAction data) {
+    switch (data.type) {
+      case NavigationType.pushNamed:
+        context
+            .goToNamed(data.routeName!, arguments: data.arguments)
+            .then((r) => data.onResult?.call(r));
+        break;
+      case NavigationType.pushNamedAndRemoveUntil:
+        context
+            .goToNamedAndRemoveUntil(
+              data.routeName!,
+              data.predicate!,
+              arguments: data.arguments,
+            )
+            .then((r) => data.onResult?.call(r));
+        break;
+      case NavigationType.pushReplacementNamed:
+        context
+            .goToNamedReplacement(data.routeName!, arguments: data.arguments)
+            .then((r) => data.onResult?.call(r));
+        break;
+      case NavigationType.push:
+        context
+            .goTo(
+              data.builder!,
+              settings: data.settings,
+              fullscreenDialog: data.fullscreenDialog,
+            )
+            .then((r) => data.onResult?.call(r));
+        break;
+      case NavigationType.pushReplacement:
+        context
+            .goToReplacement(
+              data.builder!,
+              settings: data.settings,
+              fullscreenDialog: data.fullscreenDialog,
+            )
+            .then((r) => data.onResult?.call(r));
+        break;
+      case NavigationType.pushAndRemoveUntil:
+        context
+            .goToAndRemoveUntil(
+              data.builder!,
+              data.predicate!,
+              settings: data.settings,
+              fullscreenDialog: data.fullscreenDialog,
+            )
+            .then((r) => data.onResult?.call(r));
+        break;
+      case NavigationType.pop:
+        context.pop(data.result);
+        break;
+    }
+  }
 }