+++ test/widgets/cube_widget/cube_widget_robot.dart
@@ -47,7 +47,7 @@ class CubeWidgetRobot {
   void _setupInjections({bool useMock = true}) {
     _cubeMock = useMock ? MockCubeTest() : CubeExample();
     Cubes.resetInjector();
-    Cubes.registerDependency((injector) => _cubeMock);
+    Cubes.putDependency((injector) => _cubeMock);
   }
 
   Future assetCubeReturnedInBuilder() async {